Initialize the config before express app
