+++ .eslintrc.yml
@@ -8,7 +8,7 @@ parserOptions:
   sourceType: module
 
 extends: 'eslint:recommended'
-
+parser: babel-eslint
 rules:
   no-var: 2
   key-spacing: 2
